server/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mock-alchemy==0.2.6
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.2
python-dotenv==1.0.1
regex==2024.7.24
ruff==0.5.6
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.3
